The Diaspora class is now productive
